Maximum likelihood estimators for the Principal Fitted Components (PFC) Model

Computes the ML estimators for the PFC model

MLE(X,Fy,d)

Arguments

X
vector of response variables in the inverse model, \(n x p\) matrix, each row is a response vector

Fy
vector of covariates in the inverse problem, vector containing functions of the response variable in the original multiple regression problem. Is a \(n x r\) matrix, each row is the corresponding response vector

d
number indicating the reduction subspace dimension

Details

We consider the Principal Fitted Components (PFC) model given by \(X = \mu + \Gamma\betaf(y) +\Delta\)1/2\(\epsilon\), where the variables are

\(y\) is an observed response variable of the original model. \(f\) is a known vector valued function, that takes values in Rr

\(X\) is the correspondent p x 1 observed covariates vector

\(\epsilon\) is unobserved p dimensional vector, \(\Delta\)1/2\(\epsilon\) is the error vector

and the unknown parameters (to be estimated) are

\(\mu\) a p x 1 vector of intercepts

\(\Gamma\) is a full-rank p x d matrix whose columns span the dimension reduction subspace

\(\beta\) is a full-rank d x r matrix

cov(\(\epsilon\)) = \(\Delta\), is a p x p positive definite matrix

Both coefficient matrices \(\Gamma\) and \(\beta\) are not unique, but their product p x r matrix is unique, with rank d \(\le min(p,r)\). The notation refers to Cook and Forzani (2008).

Value

List with the following components

mu

MLE estimation of the term \(\mu\) in the PFC model

beta

MLE estimation of the parameter \(\beta\) in the PFC model

gamma

MLE estimation of the parameter \(\Gamma\) in the PFC model

delta

MLE estimation of the covariance matrix \(\Delta\) in the PFC model

References

Cook, R. D. and Forzani, L. (2008). Principal Fitted components for dimension reduction in regression. Statistical Science, 23(4):485-501.

Examples

p=10
n=200
mutrue=rep(0,p)
gamatrue=as.matrix(c(1,rep(0,p-1)))
betatrue=t(as.matrix(c(1,1)))

data_sim=generate(p,n,mutrue,gamatrue,betatrue,sigmatrue=1)
Fy=data_sim$Fy
X=data_sim$X
MLE(X,Fy,d=1)

#> $mu
#>               [,1]
#>  [1,] -0.386915109
#>  [2,] -0.005455151
#>  [3,]  0.080674931
#>  [4,]  0.149416645
#>  [5,] -0.155483960
#>  [6,]  0.065566515
#>  [7,] -0.035582895
#>  [8,] -0.067898516
#>  [9,]  0.070541987
#> [10,]  0.055343925
#> 
#> $beta
#>          [,1]      [,2]
#> [1,] 1.613029 0.9896072
#> 
#> $gamma
#>                [,1]
#>  [1,]  0.9062473086
#>  [2,]  0.0043693984
#>  [3,] -0.0022197577
#>  [4,] -0.0084626687
#>  [5,]  0.0006400118
#>  [6,] -0.0078993474
#>  [7,]  0.0013556167
#>  [8,] -0.0044431204
#>  [9,] -0.0072377339
#> [10,] -0.0014514545
#> 
#> $delta
#>               [,1]         [,2]        [,3]        [,4]        [,5]
#>  [1,]  0.859291480 -0.060957870  0.01926682 -0.03123792 -0.09818991
#>  [2,] -0.060957870  0.931105530 -0.00794794  0.05560378  0.03502754
#>  [3,]  0.019266818 -0.007947940  0.90187711  0.08464409  0.10149200
#>  [4,] -0.031237923  0.055603777  0.08464409  0.82278117  0.09736358
#>  [5,] -0.098189913  0.035027541  0.10149200  0.09736358  0.97922967
#>  [6,]  0.026857002  0.017771872 -0.09736280  0.05534137 -0.04541982
#>  [7,] -0.048017592 -0.039532881 -0.01447534 -0.09163427 -0.04497941
#>  [8,]  0.102669746  0.041190755  0.06657439 -0.07255920 -0.03728341
#>  [9,]  0.003214433  0.005061948 -0.11356561  0.05590304  0.09922575
#> [10,] -0.086513178 -0.066619509 -0.04242245 -0.10581391  0.08054624
#>               [,6]        [,7]         [,8]         [,9]       [,10]
#>  [1,]  0.026857002 -0.04801759  0.102669746  0.003214433 -0.08651318
#>  [2,]  0.017771872 -0.03953288  0.041190755  0.005061948 -0.06661951
#>  [3,] -0.097362795 -0.01447534  0.066574390 -0.113565609 -0.04242245
#>  [4,]  0.055341374 -0.09163427 -0.072559198  0.055903039 -0.10581391
#>  [5,] -0.045419820 -0.04497941 -0.037283409  0.099225750  0.08054624
#>  [6,]  1.020862224  0.08133097 -0.002449878  0.093072513 -0.02259914
#>  [7,]  0.081330967  0.91940901 -0.083409375  0.034813682  0.04635996
#>  [8,] -0.002449878 -0.08340938  0.945947593  0.110829472  0.06679265
#>  [9,]  0.093072513  0.03481368  0.110829472  1.044779803  0.09807687
#> [10,] -0.022599145  0.04635996  0.066792651  0.098076873  1.06128029
#>