Data simulation

Generates samples following the Principal Fitted Components model with \(r = 2\), \(d = 1\) and covariance matrix \(\sigma\)2 times the \(p x p\) identity matrix (for more details see Bergesio et al. (2020)). The simulated model has covariate \(Fy = (y,\)y2)t, where \(y\) has uniform distribution in the \((0,4)\) interval.

generate(p, n, mutrue, gammatrue, betatrue, sigmatrue)

Arguments

p
dimension of the response vector in the inverse model

n
size of the sample required

mutrue
vector of dimension \(p\) with the true value of parameter \(\mu\)

gammatrue
matrix \(p x d\) with the true value of parameter \(\Gamma\)

betatrue
matrix \(d x r\), with \(d=1\), containing the true value of the parameter \(\beta\)

sigmatrue
constant standard deviation of the error term

Value

List with the following components

X

\(n x p\) matrix, each row is a response vector

Fy

\(n x r\) matrix, each row is the corresponding covariate vector

References

Bergesio, A., Szretter Noste, M. E. and Yohai, V. J. (2020). A robust proposal of estimation for the sufficient dimension reduction problem

Examples

require(MASS)

#> Loading required package: MASS

p=10
generate(p,n=1,mutrue=rep(0,p),gammatrue=as.matrix(c(1,rep(0,p-1))),
betatrue=t(as.matrix(c(1,1))),sigmatrue=1)

#> $Fy
#>          [,1]   [,2]
#> [1,] 1.963466 3.8552
#> 
#> $X
#>          [,1]       [,2]     [,3]     [,4]      [,5]     [,6]       [,7]
#> [1,] 5.200149 -0.6467495 0.451478 -0.33011 0.1553845 1.640898 -0.7331763
#>             [,8]      [,9]     [,10]
#> [1,] 0.007810282 0.6904836 0.3864392
#>