Initial coefficients and covariance estimators for PFC model

Compute initial estimators for the coefficient matrix and the covariance matrix. Both are strongly robust to the presence of outliers in the sample but possibly inefficient. They are a good starting point for tau-estimators for PFC model.

initial(X, Fy, aux, efficiency = 0.85)

Arguments

X
vector of response variables in the inverse model, \(n x p\) matrix, each row is a response vector

Fy
vector of covariates in the inverse problem, vector containing functions of the response variable in the original multiple regression problem. Is a \(n x p\) matrix, each row is the corresponding response vector

aux
list containing the constants for the tau-scale (with components c1, k1, c2, k2, as described in kappa_and_c)

efficiency
required efficiency for initial robust estimator

Details

This function computes the coefficient matrix concatenating the coefficients resulting from the fitting of univariate regressions using a robust estimator with high breakdown point and high efficiency (require "robust"). Univariate estimators are computed using lmRob "lmRob" function. It computes a robust estimator of errors covariance matrix (require "rrcov").

Value

The estimation is given as a list with components

beta0

coefficient matrix

delta0

S covariance matrix of residuals

See also

Examples

p=10
mutrue=rep(0,p)
gamatrue=as.matrix(c(1,rep(0,p-1)))
betatrue=t(as.matrix(c(1,1)))

data_sim=generate(p,n=200,mutrue,gamatrue,betatrue,sigmatrue=1)

aux=kappa_and_c(p,efficiency=0.85)
initial(data_sim$X,data_sim$Fy,aux,efficiency=0.85)

#> $beta0
#>              [,1]         [,2]         [,3]
#>  [1,] -0.07545820  1.236930387  0.932612434
#>  [2,]  0.09159632 -0.002129832 -0.014510932
#>  [3,]  0.02939874 -0.076097885  0.028233235
#>  [4,]  0.48227567 -0.607202284  0.136083197
#>  [5,]  0.01435626  0.020433950 -0.024474998
#>  [6,] -0.32021399  0.375870760 -0.063801457
#>  [7,] -0.26743938  0.315515045 -0.065189319
#>  [8,] -0.24232885  0.251647367 -0.046310624
#>  [9,] -0.32164760  0.193513367 -0.023749200
#> [10,]  0.08909523 -0.042652668 -0.003927533
#> 
#> $delta0
#>              [,1]          [,2]          [,3]          [,4]         [,5]
#>  [1,]  0.98019809  0.0359874961  0.0286611723 -0.0771057946 -0.125123839
#>  [2,]  0.03598750  1.0147448520  0.0682717783 -0.0050458661 -0.094050536
#>  [3,]  0.02866117  0.0682717783  0.9421324579 -0.0008629448 -0.006328672
#>  [4,] -0.07710579 -0.0050458661 -0.0008629448  1.1688360761  0.009587066
#>  [5,] -0.12512384 -0.0940505359 -0.0063286722  0.0095870656  1.118850632
#>  [6,] -0.07451806  0.0410632195  0.0305856073 -0.0669929721 -0.121122501
#>  [7,]  0.12225097  0.0973080058  0.0358317375 -0.0636431219 -0.093537678
#>  [8,] -0.19258948 -0.0097150755  0.0094974025 -0.0225974007 -0.025678978
#>  [9,] -0.07994959 -0.0006985061 -0.0311627378 -0.0107939039  0.083937298
#> [10,] -0.09886478  0.0160550718 -0.0933971551  0.1175286300 -0.085419906
#>               [,6]        [,7]         [,8]          [,9]        [,10]
#>  [1,] -0.074518056  0.12225097 -0.192589481 -0.0799495873 -0.098864777
#>  [2,]  0.041063220  0.09730801 -0.009715075 -0.0006985061  0.016055072
#>  [3,]  0.030585607  0.03583174  0.009497403 -0.0311627378 -0.093397155
#>  [4,] -0.066992972 -0.06364312 -0.022597401 -0.0107939039  0.117528630
#>  [5,] -0.121122501 -0.09353768 -0.025678978  0.0839372982 -0.085419906
#>  [6,]  0.796628840  0.07131054 -0.029249920 -0.0014254089  0.149911922
#>  [7,]  0.071310537  1.09345442 -0.120894064  0.1168935709  0.010947882
#>  [8,] -0.029249920 -0.12089406  0.939712239  0.0173921549 -0.001945888
#>  [9,] -0.001425409  0.11689357  0.017392155  1.0371615899  0.003445914
#> [10,]  0.149911922  0.01094788 -0.001945888  0.0034459141  1.105870679
#>