Tau estimate for the Principal Fitted Components Model (PFC)

Computes the tau estimators for the PFC model

tauestimate(X, Fy, d, aux, inic)

Arguments

X
vector of response variables in the inverse model, \(n x p\) matrix, each row is a response vector

Fy
vector of covariates in the inverse problem, vector containing functions of the response variable in the original multiple regression problem. Is a \(n x r\) matrix, each row is the corresponding response vector

d
number indicating the reduction subspace dimension

aux
list containing the constants for the tau-scale

inic
list containing initial estimators for the coefficient matrix and the covariance matrix

Details

We consider the Principal Fitted Components (PFC) model given by \(X = \mu + \Gamma\betaf(y) +\Delta\)1/2\(\epsilon\) where the variables are

\(y\) is an observed response variable of the original model. \(f\) is a known vector valued function, that takes values in Rr

\(X\) is the correspondent p x 1 observed covariates vector

\(\epsilon\) is unobserved p dimensional vector, \(\Delta\)1/2\(\epsilon\) is the error vector

and the unknown parameters (to be estimated) are

\(\mu\) a p x 1 vector of intercepts

\(\Gamma\) is a full-rank p x d matrix whose columns span the dimension reduction subspace

\(\beta\) is a full-rank d x r matrix

cov(\(\epsilon\)) = \(\Delta\), is a p x p positive definite matrix

Both coefficient matrices \(\Gamma\) and \(\beta\) are not unique, but their product p x r matrix is unique, with rank d \(\le min(p,r)\). The notation refers to Cook and Forzani (2008).

Value

List with the following components

mu

tau-estimator for the term \(\mu\) in the PFC model

beta

tau-estimator for the parameter \(\beta\) in the PFC model

gamma

tau-estimator for the parameter \(\Gamma\) in the PFC model

delta

tau-estimator for the covariance matrix \(\Delta\) in the PFC model

References

Bergesio, A., Szretter Noste, M. E. and Yohai, V. J. (2020). A robust proposal of estimation for the sufficient dimension reduction problem

Cook, R. D. and Forzani, L. (2008). Principal Fitted components for dimension reduction in regression. Statistical Science, 23(4):485-501.

Examples

p=10
n=200
mutrue=rep(0,p)
gamatrue=as.matrix(c(1,rep(0,p-1)))
betatrue=t(as.matrix(c(1,1)))

# to generate a dataset of n observations following the
# above PFC model, r=2
data_sim=generate(p,n,mutrue,gamatrue,betatrue,sigmatrue=1)
Fy=data_sim$Fy
X=data_sim$X

aux=kappa_and_c(p,efficiency=.9)
inic=initial(X,Fy,aux,efficiency=0.85)
tauestimate(X,Fy,d=1,aux,inic)

#> $mu
#>               [,1]
#>  [1,] -0.243511154
#>  [2,]  0.007365935
#>  [3,] -0.200036468
#>  [4,] -0.093890646
#>  [5,] -0.134315877
#>  [6,]  0.030943733
#>  [7,]  0.074739821
#>  [8,] -0.135978958
#>  [9,] -0.027015493
#> [10,] -0.017969105
#> 
#> $beta
#>          [,1]     [,2]
#> [1,] 1.547598 1.041942
#> 
#> $gamma
#>  [1]  0.8843475901 -0.0065011529  0.0154540745 -0.0008476076  0.0143403897
#>  [6] -0.0089488234 -0.0051934122  0.0131187660 -0.0070593616  0.0081187554
#> 
#> $delta
#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,]  0.85722126  0.08886269 -0.08602559  0.20562160 -0.07067790 -0.06341411
#>  [2,]  0.08886269  1.00504058 -0.15587777  0.07747573  0.11216468 -0.06147560
#>  [3,] -0.08602559 -0.15587777  1.13532471 -0.02428829 -0.08094360 -0.02444700
#>  [4,]  0.20562160  0.07747573 -0.02428829  0.98825636 -0.01437046 -0.02647680
#>  [5,] -0.07067790  0.11216468 -0.08094360 -0.01437046  0.94346170  0.03741923
#>  [6,] -0.06341411 -0.06147560 -0.02444700 -0.02647680  0.03741923  0.95824002
#>  [7,] -0.04152928 -0.09162647  0.02903779  0.18310443 -0.14848678 -0.12948328
#>  [8,]  0.05703462  0.05696580  0.01562002  0.16128339  0.04806727 -0.02803199
#>  [9,]  0.01748799 -0.10981577  0.22372711  0.01997940 -0.13777782  0.10564786
#> [10,]  0.04809085  0.07390956 -0.11027696  0.09451212 -0.13027690  0.04562906
#>              [,7]        [,8]        [,9]       [,10]
#>  [1,] -0.04152928  0.05703462  0.01748799  0.04809085
#>  [2,] -0.09162647  0.05696580 -0.10981577  0.07390956
#>  [3,]  0.02903779  0.01562002  0.22372711 -0.11027696
#>  [4,]  0.18310443  0.16128339  0.01997940  0.09451212
#>  [5,] -0.14848678  0.04806727 -0.13777782 -0.13027690
#>  [6,] -0.12948328 -0.02803199  0.10564786  0.04562906
#>  [7,]  1.20949379 -0.05146638  0.03213792  0.02045262
#>  [8,] -0.05146638  0.88917606  0.09836186 -0.05668995
#>  [9,]  0.03213792  0.09836186  0.90890435  0.11477070
#> [10,]  0.02045262 -0.05668995  0.11477070  0.88562579
#>